+++ tests/test_correlatednoise.py
@@ -36,37 +36,12 @@ rseed=12345
 smallim_size = 16 # size of image when we test correlated noise properties using small inputs
 smallim_size_odd = 17 # odd-sized version of the above for odd/even relevant tests (e.g. draw)
 largeim_size = 12 * smallim_size # ditto, but when we need a larger image
-xlargeim_size =long(np.ceil(1.41421356 * largeim_size)) # sometimes, for precision tests, we 
-                                                        # need a very large image that will 
-                                                        # fit a large image within it, even if 
-                                                        # rotated
-
-# Decimals for comparison (one for fine detail, another for comparing stochastic quantities)
-decimal_approx = 2
-decimal_precise = 7
 
 # Number of positions to test in nonzero lag uncorrelated tests
 npos_test = 3
 
 # Number of CorrelatedNoises to sum over to get slightly better statistics for noise generation test
 nsum_test = 7
-
-# Set up some variables for testing of noise symmetry.  If we run this test suite directly, do a
-# more stringent test that takes more time, otherwise do a looser test that takes less time.
-if __name__ == "__main__":
-    symm_divide = 2. # test to precision determined by decimal_approx (usually 2%)
-    symm_size_mult = 6 # make really huge images
-    # The convolve_cosmos test, which includes a lot of the correlated noise functionality is
-    # fairly sensitive at 2dp, but takes ~200s on a mid-range laptop
-    decimal_convolve_cosmos = 2
-    nsum_test_convolve_cosmos = 1000
-else:
-    symm_divide = 3. # test to worse precision than determined by decimal_approx (usually 3%)
-    symm_size_mult = 3 # make only moderately huge images
-    # Basic settings for convolve_cosmos, will only catch basic screwups
-    decimal_convolve_cosmos = 1
-    nsum_test_convolve_cosmos = 10
-
 
 def setup_uncorrelated_noise(deviate, size):
     """Makes and returns uncorrelated noise fields for later use in generating derived correlated
@@ -99,19 +74,29 @@ def test_uncorrelated_noise_zero_lag():
     t1 = time.time()
     sigmas = [3.e-9, 49., 1.11e11]  # some wide ranging sigma values for the noise field
     # loop through the sigmas
-    cf_zero = 0.
     gd = galsim.GaussianDeviate(rseed)
     for sigma in sigmas:
         # Test the estimated value is good to 1% of the input variance; we expect this!
         # Note we make multiple correlation funcs and average their zero lag to beat down noise
+        cf_zero = 0.
         for i in range(nsum_test):
             uncorr_noise_image = setup_uncorrelated_noise(gd, largeim_size) * sigma
-            cn = galsim.CorrelatedNoise(uncorr_noise_image, gd, scale=1.)
+            cn = galsim.CorrelatedNoise(uncorr_noise_image, gd)
             cf_zero += cn._profile.xValue(galsim.PositionD(0., 0.))
         cf_zero /= float(nsum_test)
         np.testing.assert_almost_equal(
-            cf_zero / sigma**2, 1., decimal=decimal_approx,
+            cf_zero / sigma**2, 1., decimal=2,
             err_msg="Zero distance noise correlation value does not match input noise variance.")
+
+        # Repeat using UncorrelatedNoise
+        ucn = galsim.UncorrelatedNoise(variance=sigma**2)
+        ucf_zero = ucn._profile.xValue(galsim.PositionD(0.,0.))
+        np.testing.assert_almost_equal(
+            ucf_zero / sigma**2, 1., decimal=5,
+            err_msg="Zero distance noise correlation value does not match variance value " + 
+            "provided to UncorrelatedNoise.")
+
+
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
 
@@ -120,30 +105,104 @@ def test_uncorrelated_noise_nonzero_lag():
     randomly chosen positions.
     """
     t1 = time.time()
+    ud = galsim.UniformDeviate(rseed)
+    sigma = 1.7  # Don't use sigma = 1, to notice sqrt error.
+    gn = galsim.GaussianNoise(ud, sigma=sigma)
     # Set up some random positions (within and outside) the bounds of the table inside the
     # CorrelatedNoise then test
-    uncorr_noise_image = galsim.ImageD(largeim_size, largeim_size)
-    ud = galsim.UniformDeviate(rseed)
-    gn = galsim.GaussianNoise(ud, sigma=1.)
-    for i in range(npos_test):
+    for rpos in [ 1.3, 23., 84., 243. ]:
         # Note we make multiple noise fields and correlation funcs and average non-zero lag values
         # to beat down noise
         cf_test_value = 0.
         for i in range(nsum_test):
-            uncorr_noise_image.addNoise(gn)
-            cn = galsim.CorrelatedNoise(uncorr_noise_image, ud, scale=1.)
-            # generate the test position at least one pixel away from the origin
-            rpos = 2. + ud() * (largeim_size - 2.) # this can go outside table bounds
-            tpos = 2. * np.pi * ud()
+            tpos = 2. * np.pi * ud()  # A new random direction each time
             pos = galsim.PositionD(rpos * np.cos(tpos), rpos * np.sin(tpos))
+            uncorr_noise_image = setup_uncorrelated_noise(ud, largeim_size) * sigma
+            cn = galsim.CorrelatedNoise(uncorr_noise_image, ud)
+            # generate the test position at least one pixel away from the origin
             cf_test_value += cn._profile.xValue(pos)
-            uncorr_noise_image.setZero()
         cf_test_value /= float(nsum_test)
         # Then test this estimated value is good to within our chosen decimal place of zero
         np.testing.assert_almost_equal(
-            cf_test_value, 0., decimal=decimal_approx,
+            cf_test_value, 0., decimal=2,
             err_msg="Non-zero distance noise correlation value not sufficiently close to target "+
             "value of zero.")
+
+        # Repeat using UncorrelatedNoise
+        ucn = galsim.UncorrelatedNoise(variance=sigma**2)
+        ucf_test_value = ucn._profile.xValue(pos)
+        np.testing.assert_almost_equal(
+            ucf_test_value, 0., decimal=5,
+            err_msg="Non-zero distance noise correlation value not sufficiently close to target "+
+            "value of zero.")
+
+    t2 = time.time()
+    print 'time for %s = %.2f'%(funcname(), t2 - t1)
+
+def test_uncorrelated_noise_output():
+    """Test that noise generated by a UncorrelatedNoise has the expected variance.
+    """
+    import math
+    t1 = time.time()
+
+    ud = galsim.UniformDeviate(rseed)
+    sigma = 1.7
+
+    # Note: Test non-square and odd size in y.
+    im = galsim.ImageD(2*largeim_size, 3*largeim_size+1, init_value=0.)
+
+    # If no wcs is given, then the variance is the pixel variance
+    ucn = galsim.UncorrelatedNoise(rng=ud, variance=sigma**2)
+    im.addNoise(ucn)
+    im_var = im.array.var()
+    np.testing.assert_array_almost_equal(
+        im_var/sigma**2, 1.0, decimal=2,
+        err_msg="Generated uncorrelated noise field (2) does not have expected variance")
+
+    # If we provide a wcs, then the wcs given to UncorrelatedNoise should match the wcs
+    # of the image.  In this case, the resulting pixel variance should still match the 
+    # provided variance parameter.
+    if False:
+        # The last one is a sufficient unit test, but if it fails, it is helpful to
+        # test the simpler ones first.
+        wcs_list = [ galsim.PixelScale(1.0),
+                     galsim.PixelScale(0.23),
+                     galsim.JacobianWCS(0.23, 0.01, -0.05, 0.21),
+                     galsim.AffineTransform(0.23, 0.31, -0.15, 0.21, galsim.PositionD(73,45)) ]
+    else:
+        wcs_list = [ galsim.AffineTransform(0.23, 0.31, -0.15, 0.21, galsim.PositionD(73,45)) ]
+    for wcs in wcs_list:
+        scale = math.sqrt(wcs.pixelArea())
+        im.wcs = wcs
+        print 'wcs = ',wcs
+        print 'scale = ',scale
+
+        # First test with an UncorrelatedNoise object
+        ucn = galsim.UncorrelatedNoise(rng=ud, variance=sigma**2, wcs=wcs)
+        im.setZero()
+        im.addNoise(ucn)
+        im_var = im.array.var()
+        print 'im_mean = ',im.array.mean()
+        print 'im_std = ',im.array.std()
+        print 'im_var = ',im_var
+        np.testing.assert_array_almost_equal(
+            im_var/sigma**2, 1.0, decimal=2,
+            err_msg="Generated uncorrelated noise field (1) does not have expected variance")
+
+        # Test with a regular CorrelatedNoise object built from an image with this variance
+        uncorr_noise_image = setup_uncorrelated_noise(ud, 10*largeim_size) * sigma
+        uncorr_noise_image.wcs = wcs
+        cn = galsim.CorrelatedNoise(uncorr_noise_image, ud)
+        im.setZero()
+        im.addNoise(cn)
+        im_var = im.array.var()
+        print 'im_mean = ',im.array.mean()
+        print 'im_std = ',im.array.std()
+        print 'im_var = ',im_var
+        np.testing.assert_array_almost_equal(
+            im_var/sigma**2, 1.0, decimal=2,
+            err_msg="Generated uncorrelated noise field (2) does not have expected variance")
+
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
 
@@ -155,7 +214,7 @@ def test_uncorrelated_noise_symmetry_90degree_rotation():
     t1 = time.time()
     ud = galsim.UniformDeviate(rseed)
     uncorr_noise_small = setup_uncorrelated_noise(ud, smallim_size)
-    cn = galsim.CorrelatedNoise(uncorr_noise_small, ud, scale=1.) # small image is fine here
+    cn = galsim.CorrelatedNoise(uncorr_noise_small, ud) # small image is fine here
     # Set up some random positions (within and outside) the bounds of the table inside the corrfunc
     # then test
     for i in range(npos_test):
@@ -165,8 +224,7 @@ def test_uncorrelated_noise_symmetry_90degree_rotation():
         cf_test1 = cn._profile.xValue(pos)
         cf_test2 = cn._profile.xValue(-pos)
         np.testing.assert_almost_equal(
-            cf_test1, cf_test2,
-            decimal=decimal_precise,
+            cf_test1, cf_test2, decimal=7,
             err_msg="Non-zero distance noise correlation values not two-fold rotationally "+
             "symmetric.")
     # Then test that CorrelatedNoise rotation methods produces the same output as initializing 
@@ -177,7 +235,7 @@ def test_uncorrelated_noise_symmetry_90degree_rotation():
     for i, angle in zip(range(len(angles)), angles):
         noise_ref = galsim.ImageD(
             np.ascontiguousarray(np.rot90(uncorr_noise_small.array, k=i+1)))
-        cn_ref = galsim.CorrelatedNoise(noise_ref, ud, scale=1.)
+        cn_ref = galsim.CorrelatedNoise(noise_ref, ud)
         cn_test1 = cn.rotate(angle)
         # Then check some positions inside the bounds of the original image
         for i in range(npos_test):
@@ -188,7 +246,7 @@ def test_uncorrelated_noise_symmetry_90degree_rotation():
             cf_test1 = cn_test1._profile.xValue(pos)
             # Then test these estimated value is good to within our chosen decimal place
             np.testing.assert_almost_equal(
-                cf_test1, cf_ref, decimal=decimal_precise,
+                cf_test1, cf_ref, decimal=7,
                 err_msg="Uncorrelated noise failed 90 degree rotate() method test.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -206,18 +264,18 @@ def test_xcorr_noise_basics_symmetry_90degree_rotation():
     for i in range(nsum_test):
         uncorr_noise = setup_uncorrelated_noise(ud, largeim_size)
         xnoise = make_xcorr_from_uncorr(uncorr_noise)
-        xcn = galsim.CorrelatedNoise(xnoise, ud, scale=1.)
+        xcn = galsim.CorrelatedNoise(xnoise, ud)
         cf_zero += xcn._profile.xValue(galsim.PositionD(0., 0.))
         cf_10 += xcn._profile.xValue(galsim.PositionD(1., 0.))
     cf_zero /= float(nsum_test)
     cf_10 /= float(nsum_test)
     # Then test the zero-lag value is good to 1% of the input variance; we expect this!
     np.testing.assert_almost_equal(
-        cf_zero, 1., decimal=decimal_approx,
+        cf_zero, 1., decimal=2,
         err_msg="Zero distance noise correlation value does not match input noise variance.")
     # Then test the (1, 0) value is good to 1% of the input variance (0.5); we expect this!
     np.testing.assert_almost_equal(
-        cf_10, .5, decimal=decimal_approx,
+        cf_10, .5, decimal=2,
         err_msg="Noise correlation value at (1, 0) does not match input covariance.")
     # Then set up some random positions (within and outside) the bounds of the table inside the 
     # corrfunc (the last one made is fine) then test for symmetry
@@ -229,7 +287,7 @@ def test_xcorr_noise_basics_symmetry_90degree_rotation():
         cf_test2 = xcn._profile.xValue(-pos)
         # Then test this estimated value is good to within our chosen decimal place of zero
         np.testing.assert_almost_equal(
-            cf_test1, cf_test2, decimal=decimal_precise, # should be good to machine precision
+            cf_test1, cf_test2, decimal=12, # should be good to machine precision
             err_msg="Non-zero distance noise correlation values not two-fold rotationally "+
             "symmetric for x correlated noise field.")
     # Then test that CorrelatedNoise rotation methods produces the same output as initializing 
@@ -240,7 +298,7 @@ def test_xcorr_noise_basics_symmetry_90degree_rotation():
     for i, angle in zip(range(len(angles)), angles):
         noise_ref = galsim.ImageD(
             np.ascontiguousarray(np.rot90(xnoise.array, k=i+1)))
-        xcn_ref = galsim.CorrelatedNoise(noise_ref, ud, scale=1.)
+        xcn_ref = galsim.CorrelatedNoise(noise_ref, ud)
         xcn_test1 = xcn.rotate(angle)
         # Then check some positions inside the bounds of the original image
         for i in range(npos_test):
@@ -251,7 +309,7 @@ def test_xcorr_noise_basics_symmetry_90degree_rotation():
             xcf_test1 = xcn_test1._profile.xValue(pos)
             # Then test these estimated value is good to within our chosen decimal place
             np.testing.assert_almost_equal(
-                xcf_test1, xcf_ref, decimal=decimal_precise,
+                xcf_test1, xcf_ref, decimal=7,
                 err_msg="x-correlated noise failed 90 degree rotate() method test.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -269,18 +327,18 @@ def test_ycorr_noise_basics_symmetry_90degree_rotation():
     for i in range(nsum_test):
         uncorr_noise = setup_uncorrelated_noise(ud, largeim_size)
         ynoise = make_ycorr_from_uncorr(uncorr_noise)
-        ycn = galsim.CorrelatedNoise(ynoise, ud, scale=1.)
+        ycn = galsim.CorrelatedNoise(ynoise, ud)
         cf_zero += ycn._profile.xValue(galsim.PositionD(0., 0.))
         cf_01 += ycn._profile.xValue(galsim.PositionD(0., 1.))
     cf_zero /= float(nsum_test)
     cf_01 /= float(nsum_test)
     # Then test the zero-lag value is good to 1% of the input variance; we expect this!
     np.testing.assert_almost_equal(
-        cf_zero, 1., decimal=decimal_approx,
+        cf_zero, 1., decimal=2,
         err_msg="Zero distance noise correlation value does not match input noise variance.")
     # Then test the (0, 1) value is good to 1% of the input variance (0.5); we expect this!
     np.testing.assert_almost_equal(
-        cf_01, .5, decimal=decimal_approx,
+        cf_01, .5, decimal=2,
         err_msg="Noise correlation value at (0, 1) does not match input covariance.")
     # Then set up some random positions (within and outside) the bounds of the table inside the 
     # corrfunc (the last one made is fine) then test for symmetry
@@ -292,7 +350,7 @@ def test_ycorr_noise_basics_symmetry_90degree_rotation():
         cf_test2 = ycn._profile.xValue(-pos)
         # Then test this estimated value is good to within our chosen decimal place of zero
         np.testing.assert_almost_equal(
-            cf_test1, cf_test2, decimal=decimal_precise, # should be good to machine precision
+            cf_test1, cf_test2, decimal=12, # should be good to machine precision
             err_msg="Non-zero distance noise correlation values not two-fold rotationally "+
             "symmetric for x correlated noise field.")
     # Then test that CorrelatedNoise rotation methods produces the same output as initializing 
@@ -303,7 +361,7 @@ def test_ycorr_noise_basics_symmetry_90degree_rotation():
     for i, angle in zip(range(len(angles)), angles):
         noise_ref = galsim.ImageD(
             np.ascontiguousarray(np.rot90(ynoise.array, k=i+1)))
-        ycn_ref = galsim.CorrelatedNoise(noise_ref, ud, scale=1.)
+        ycn_ref = galsim.CorrelatedNoise(noise_ref, ud)
         ycn_test1 = ycn.rotate(angle)
         # Then check some positions inside the bounds of the original image
         for i in range(npos_test):
@@ -314,7 +372,7 @@ def test_ycorr_noise_basics_symmetry_90degree_rotation():
             ycf_test1 = ycn_test1._profile.xValue(pos)
             # Then test these estimated value is good to within our chosen decimal place
             np.testing.assert_almost_equal(
-                ycf_test1, ycf_ref, decimal=decimal_precise,
+                ycf_test1, ycf_ref, decimal=7,
                 err_msg="y-correlated noise failed 90 degree rotate() method test.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -327,7 +385,7 @@ def test_arbitrary_rotation():
     t1 = time.time()
     ud = galsim.UniformDeviate(rseed)
     ynoise_small = make_ycorr_from_uncorr(setup_uncorrelated_noise(ud, smallim_size))
-    cn = galsim.CorrelatedNoise(ynoise_small, ud, scale=1.) # use something >0
+    cn = galsim.CorrelatedNoise(ynoise_small, ud)
     for i in range(npos_test):
         rot_angle = 2. * np.pi * ud()
         rpos = ud() * smallim_size # look in the vicinity of the action near the centre
@@ -340,8 +398,7 @@ def test_arbitrary_rotation():
         # then create rotated cns for comparison
         cn_rot1 = cn.rotate(rot_angle * galsim.radians)
         np.testing.assert_almost_equal(
-            cn._profile.xValue(pos_rot), cn_rot1._profile.xValue(pos_ref), 
-            decimal=decimal_precise, # this should be good at very high accuracy 
+            cn._profile.xValue(pos_rot), cn_rot1._profile.xValue(pos_ref), decimal=12,
             err_msg="Noise correlated in the y direction failed rotate() "+
             "method test for arbitrary rotations.")
     t2 = time.time()
@@ -355,7 +412,7 @@ def test_scaling():
     t1 = time.time()
     ud = galsim.UniformDeviate(rseed)
     xnoise_small = make_xcorr_from_uncorr(setup_uncorrelated_noise(ud, smallim_size))
-    cn = galsim.CorrelatedNoise(xnoise_small, ud, scale=1.)
+    cn = galsim.CorrelatedNoise(xnoise_small, ud)
     scalings = [7.e-13, 424., 7.9e23]
     for scale in scalings:
        cn_test1 = cn.expand(scale)
@@ -364,8 +421,7 @@ def test_scaling():
            tpos = 2. * np.pi * ud()
            pos_ref = galsim.PositionD(rpos * np.cos(tpos), rpos * np.sin(tpos))
            np.testing.assert_almost_equal(
-               cn_test1._profile.xValue(pos_ref), cn._profile.xValue(pos_ref / scale),
-               decimal=decimal_precise,
+               cn_test1._profile.xValue(pos_ref), cn._profile.xValue(pos_ref / scale), decimal=7,
                err_msg="Noise correlated in the y direction failed expand() scaling test.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -377,7 +433,7 @@ def test_jacobian():
     t1 = time.time()
     ud = galsim.UniformDeviate(rseed)
     ynoise_small = make_ycorr_from_uncorr(setup_uncorrelated_noise(ud, smallim_size))
-    cn = galsim.CorrelatedNoise(ynoise_small, ud, scale=1.)
+    cn = galsim.CorrelatedNoise(ynoise_small, ud)
     dudx = 0.241
     dudy = 0.051
     dvdx = -0.098
@@ -390,8 +446,7 @@ def test_jacobian():
         pos_test = galsim.PositionD(pos_ref.x * dudx + pos_ref.y * dudy,
                                     pos_ref.x * dvdx + pos_ref.y * dvdy)
         np.testing.assert_almost_equal(
-            cn_test1._profile.xValue(pos_test), cn._profile.xValue(pos_ref),
-            decimal=decimal_precise,
+            cn_test1._profile.xValue(pos_test), cn._profile.xValue(pos_ref), decimal=7,
             err_msg="Noise correlated in the y direction failed transform() test")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -418,12 +473,12 @@ def test_drawImage():
     cf_array = utilities.roll2d(cf_array, (cf_array.shape[0] / 2, cf_array.shape[1] / 2))
     # Then use the CorrelatedNoise class for comparison (don't use periodicity correction for
     # comparison with naive results above)
-    cn = galsim.CorrelatedNoise(uncorr_noise_small_odd, gd, scale=1., correct_periodicity=False)
+    cn = galsim.CorrelatedNoise(uncorr_noise_small_odd, gd, correct_periodicity=False)
     testim1 = galsim.ImageD(smallim_size_odd, smallim_size_odd)
     cn.drawImage(testim1, scale=1.)
     # Then compare the odd-sized arrays:
     np.testing.assert_array_almost_equal(
-        testim1.array, cf_array, decimal=decimal_precise, 
+        testim1.array, cf_array, decimal=7,
         err_msg="Drawn image (odd-sized) does not match independently calculated correlated noise.")
     # Now we do even
     uncorr_noise_small = setup_uncorrelated_noise(gd, smallim_size)
@@ -434,12 +489,12 @@ def test_drawImage():
     cf_array = (np.fft.ifft2(ps_array)).real / float(np.product(np.shape(ft_array)))
     cf_array = utilities.roll2d(cf_array, (cf_array.shape[0] / 2, cf_array.shape[1] / 2))
     # Then use the CorrelatedNoise class for comparison (as above don't correct for periodicity)
-    cn = galsim.CorrelatedNoise(uncorr_noise_small, gd, scale=1., correct_periodicity=False)
+    cn = galsim.CorrelatedNoise(uncorr_noise_small, gd, correct_periodicity=False)
     testim1 = galsim.ImageD(smallim_size, smallim_size)
     cn.drawImage(testim1, scale=1.)
     # Then compare the even-sized arrays:
     np.testing.assert_array_almost_equal(
-        testim1.array, cf_array, decimal=decimal_precise, 
+        testim1.array, cf_array, decimal=7,
         err_msg="Drawn image (even-sized) does not match independently calculated correlated "+
         "noise.")
     t2 = time.time()
@@ -454,8 +509,10 @@ def test_output_generation_basic():
     # Get the correlated noise from an image of some x-correlated noise
     xnoise_large = make_xcorr_from_uncorr(setup_uncorrelated_noise(ud, largeim_size))
     # Note the noise we generate in these tests *is* periodic, so use non-default settings
-    cn = galsim.CorrelatedNoise(xnoise_large, ud, scale=.18, correct_periodicity=False)
-    refim = galsim.ImageD(smallim_size, smallim_size)
+    xnoise_large.scale = 0.18
+    cn = galsim.CorrelatedNoise(xnoise_large, ud, correct_periodicity=False)
+    # Note non-square and y is odd
+    refim = galsim.ImageD(smallim_size, smallim_size+11)
     # Draw this for reference
     cn.drawImage(refim, scale=.18)
     # Generate a large image containing noise according to this function
@@ -463,20 +520,20 @@ def test_output_generation_basic():
     outimage.addNoise(cn)
     # Summed (average) CorrelatedNoises should be approximately equal to the input, so average
     # multiple CFs
-    cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, scale=.18, correct_periodicity=False)
+    cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, correct_periodicity=False)
     # Draw the summed CF to an image for comparison 
-    testim = galsim.ImageD(smallim_size, smallim_size)
+    testim = galsim.ImageD(smallim_size, smallim_size+11)
     cn_2ndlevel.drawImage(testim, scale=.18, add_to_image=True)
     for i in range(nsum_test - 1):
         # Then repeat
         outimage.setZero()
         outimage.addNoise(cn)
-        cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, scale=.18, correct_periodicity=False)
+        cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, correct_periodicity=False)
         cn_2ndlevel.drawImage(testim, scale=.18, add_to_image=True)
     # Then take average
     testim /= float(nsum_test)
     np.testing.assert_array_almost_equal(
-        testim.array, refim.array, decimal=decimal_approx,
+        testim.array, refim.array, decimal=2,
         err_msg="Generated noise field (basic) does not match input correlation properties.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -504,10 +561,12 @@ def test_output_generation_rotated():
     # TODO: It would be good to understand more about the detailed interpolant behaviour though...
     ud = galsim.UniformDeviate(rseed)
     # Get the correlated noise from an image of some y-correlated noise
+    xlargeim_size =long(np.ceil(1.41421356 * largeim_size)) 
+    # need a very large image that will fit a large image within it, even if rotated
     ynoise_xlarge = make_ycorr_from_uncorr(setup_uncorrelated_noise(ud, xlargeim_size))
     # Subtract the mean
     ynoise_xlarge -= ynoise_xlarge.array.mean()
-    cn = galsim.CorrelatedNoise(ynoise_xlarge, ud, scale=1., correct_periodicity=False)
+    cn = galsim.CorrelatedNoise(ynoise_xlarge, ud, correct_periodicity=False)
     # Then loop over some angles
     angles = [28.7 * galsim.degrees, 135. * galsim.degrees]
     for angle in angles:
@@ -520,20 +579,20 @@ def test_output_generation_rotated():
         outimage.addNoise(cn_rot)
         # Summed (average) CorrelatedNoises should be approximately equal to the input, so avg
         # multiple CFs
-        cn_2ndlevel = galsim.CorrelatedNoise(
-            outimage, ud, scale=1., subtract_mean=False, correct_periodicity=False)
+        cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, subtract_mean=False,
+                                             correct_periodicity=False)
         for i in range(nsum_test - 1):
             # Then repeat
             outimage.setZero()
             outimage.addNoise(cn_rot)
-            cn_2ndlevel += galsim.CorrelatedNoise(
-                outimage, ud, scale=1., subtract_mean=False, correct_periodicity=False)
+            cn_2ndlevel += galsim.CorrelatedNoise(outimage, ud, subtract_mean=False,
+                                                  correct_periodicity=False)
         cn_2ndlevel /= float(nsum_test)
         # Then draw the summed CF to an image for comparison 
         testim = galsim.ImageD(smallim_size, smallim_size)
         cn_2ndlevel.drawImage(testim, scale=1.)
         np.testing.assert_array_almost_equal(
-            testim.array, refim.array, decimal=decimal_approx,
+            testim.array, refim.array, decimal=2,
             err_msg="Generated noise field (rotated) does not match input correlation properties.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -546,7 +605,7 @@ def test_output_generation_magnified():
     # Get the correlated noise from an image of some y-correlated noise
     ynoise_large = make_ycorr_from_uncorr(setup_uncorrelated_noise(ud, largeim_size))
     # Get the correlated noise
-    cn = galsim.CorrelatedNoise(ynoise_large, ud, scale=1., correct_periodicity=False)
+    cn = galsim.CorrelatedNoise(ynoise_large, ud, correct_periodicity=False)
     refim = galsim.ImageD(smallim_size, smallim_size)
     # Draw this for reference
     cn.drawImage(refim, scale=1.)
@@ -561,6 +620,7 @@ def test_output_generation_magnified():
         outimage.addNoise(cn_scl)
         # Summed (average) CorrelatedNoises should be approximately equal to the input, so avg
         # multiple CFs
+        # Make the CorrelatedNoise object with scale=1. rather than image scale
         cn_2ndlevel = galsim.CorrelatedNoise(outimage, ud, scale=1., correct_periodicity=False)
         for i in range(nsum_test - 1): # Need to add here to nsum_test to beat down noise
             # Then repeat
@@ -573,7 +633,7 @@ def test_output_generation_magnified():
         testim = galsim.ImageD(smallim_size, smallim_size)
         cn_2ndlevel.drawImage(testim, scale=1.)
         np.testing.assert_array_almost_equal(
-            testim.array, refim.array, decimal=decimal_approx,
+            testim.array, refim.array, decimal=2,
             err_msg="Generated noise does not match (magnified) input correlation properties.")
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -620,7 +680,7 @@ def test_copy():
     # rootps_store cache is *not* copied over and when the CF is redrawn the GSObject may make
     # slight approximations.  So we'll just test at high precision:
     np.testing.assert_array_almost_equal(
-        outim1.array, outim2.array, decimal=decimal_precise,
+        outim1.array, outim2.array, decimal=7,
         err_msg="Copied correlated noise does not produce the same noise field as the parent "+
         "despite sharing the same RNG.")
     # To illustrate the point above, we'll run a test in which the rootps is not stored in cn when
@@ -632,9 +692,8 @@ def test_copy():
     cn = cn.copy(rng=cn.getRNG().duplicate())
     outim1.addNoise(cn)
     outim2.addNoise(cn_copy)
-    decimal_very_precise = 14
     np.testing.assert_array_almost_equal(
-        outim1.array, outim2.array, decimal=decimal_precise,
+        outim1.array, outim2.array, decimal=12,
         err_msg="Copied correlated noise does not produce the same noise field as the parent "+
         "despite sharing the same RNG (high precision test).")
     t2 = time.time()
@@ -650,7 +709,8 @@ def test_cosmos_and_whitening():
     cosmos_scale = 7.5 # Use some non-default, non-unity value of COSMOS pixel spacing
     ccn = galsim.getCOSMOSNoise(rng=gd, cosmos_scale=cosmos_scale)
     # large image to beat down noise
-    outimage = galsim.ImageD(3 * largeim_size, 3 * largeim_size, scale=cosmos_scale)
+    # Non-square and x size is odd.
+    outimage = galsim.ImageD(3 * largeim_size + 11, 3 * largeim_size, scale=cosmos_scale)
     outimage.addNoise(ccn)  # Add the COSMOS noise
     # Then estimate correlation function from generated noise
     cntest_correlated = galsim.CorrelatedNoise(outimage, ccn.getRNG())
@@ -660,7 +720,7 @@ def test_cosmos_and_whitening():
     cftest00 = cntest_correlated._profile.xValue(pos)
     # Test variances first
     np.testing.assert_almost_equal(
-        cftest00 / cf00, 1., decimal=decimal_approx,
+        cftest00 / cf00, 1., decimal=2,
         err_msg="Noise field generated with COSMOS CorrelatedNoise does not approximately match "+
         "input variance")
     # Then test (1, 0), (0, 1), (1,-1) and (1,1) values
@@ -670,7 +730,7 @@ def test_cosmos_and_whitening():
         cf = ccn._profile.xValue(pos)
         cftest = cntest_correlated._profile.xValue(pos)
         np.testing.assert_almost_equal(
-            cftest / cftest00, cf / cf00, decimal=decimal_approx,
+            cftest / cftest00, cf / cf00, decimal=2,
             err_msg="Noise field generated with COSMOS CorrelatedNoise does not have "+
             "approximately matching interpixel covariances")
     # Now whiten the noise field, and check that its variance and covariances are as expected
@@ -680,7 +740,7 @@ def test_cosmos_and_whitening():
     cftest00 = cntest_whitened._profile.xValue(galsim.PositionD(0., 0.))
     # Test variances first
     np.testing.assert_almost_equal(
-        cftest00 / whitened_variance, 1., decimal=decimal_approx,
+        cftest00 / whitened_variance, 1., decimal=2,
         err_msg="Noise field generated by whitening COSMOS CorrelatedNoise does not approximately "+
         "match theoretical variance")
     # Then test (1, 0), (0, 1), (1,-1) and (1,1) values
@@ -689,7 +749,7 @@ def test_cosmos_and_whitening():
         pos = galsim.PositionD(xpos, ypos)
         cftest = cntest_whitened._profile.xValue(pos)
         np.testing.assert_almost_equal(
-            cftest / cftest00, 0., decimal=decimal_approx,
+            cftest / cftest00, 0., decimal=2,
             err_msg="Noise field generated by whitening COSMOS CorrelatedNoise does not have "+
             "approximately zero interpixel covariances")
     # Now test whitening but having first expanded and sheared the COSMOS noise correlation
@@ -701,7 +761,7 @@ def test_cosmos_and_whitening():
     cftest00 = cntest_whitened._profile.xValue(galsim.PositionD(0., 0.))
     # Test variances first
     np.testing.assert_almost_equal(
-        cftest00 / wht_variance, 1., decimal=decimal_approx,
+        cftest00 / wht_variance, 1., decimal=2,
         err_msg="Noise field generated by whitening rotated, sheared, magnified COSMOS "+
         "CorrelatedNoise does not approximately match theoretical variance")
     # Then test (1, 0), (0, 1), (1,-1) and (1,1) values
@@ -710,7 +770,7 @@ def test_cosmos_and_whitening():
         pos = galsim.PositionD(xpos, ypos)
         cftest = cntest_whitened._profile.xValue(pos)
         np.testing.assert_almost_equal(
-            cftest / cftest00, 0., decimal=decimal_approx,
+            cftest / cftest00, 0., decimal=2,
             err_msg="Noise field generated by whitening rotated, sheared, magnified COSMOS "+
             "CorrelatedNoise does not have approximately zero interpixel covariances")
     # Then convolve with a ground-based PSF and pixel, generate some more correlated noise
@@ -733,7 +793,7 @@ def test_cosmos_and_whitening():
     cftest00 = cntest_whitened._profile.xValue(galsim.PositionD(0., 0.))
     # Test variances first
     np.testing.assert_almost_equal(
-        cftest00 / wht_variance, 1., decimal=decimal_approx,
+        cftest00 / wht_variance, 1., decimal=2,
         err_msg="Noise field generated by whitening rotated, sheared, magnified, convolved COSMOS "+
         "CorrelatedNoise does not approximately match theoretical variance")
     # Then test (1, 0), (0, 1), (1,-1) and (1,1) values
@@ -742,7 +802,7 @@ def test_cosmos_and_whitening():
         pos = galsim.PositionD(xpos, ypos)
         cftest = cntest_whitened._profile.xValue(pos)
         np.testing.assert_almost_equal(
-            cftest / cftest00, 0., decimal=decimal_approx,
+            cftest / cftest00, 0., decimal=2,
             err_msg="Noise field generated by whitening rotated, sheared, magnified, convolved "+
             "COSMOS CorrelatedNoise does not have approximately zero interpixel covariances")
     t2 = time.time()
@@ -754,6 +814,15 @@ def test_symmetrizing():
     convolution with a ground-based PSF.
     """
     t1 = time.time()
+
+    if __name__ == "__main__":
+        # symm_divide determines how close to zero we need to be.  Bigger is looser tolerance.
+        symm_divide = 2.
+        symm_size_mult = 6 # make really huge images
+    else:
+        symm_divide = 3.
+        symm_size_mult = 3 # make only moderately huge images
+
     gd = galsim.GaussianDeviate(rseed)
     cosmos_scale = 7.5 # Use some non-default, non-unity value of COSMOS pixel spacing
     ccn = galsim.getCOSMOSNoise(
@@ -769,9 +838,10 @@ def test_symmetrizing():
     symmetrized_variance = ccn.symmetrizeImage(outimage, order=4)
     cntest_symmetrized = galsim.CorrelatedNoise(outimage, ccn.getRNG()) # Get the correlation function
     cftest00 = cntest_symmetrized._profile.xValue(galsim.PositionD(0., 0.))
+
     # Test variances first
     np.testing.assert_almost_equal(
-        (cftest00/symmetrized_variance - 1.)/symm_divide, 0, decimal=decimal_approx,
+        (cftest00/symmetrized_variance - 1.)/symm_divide, 0, decimal=2,
         err_msg="Noise field generated by symmetrizing COSMOS CorrelatedNoise does not "+
         "approximately match theoretical variance")
     # Then test (1, 0), (0, 1), (-1,0) and (0,-1) values
@@ -783,7 +853,7 @@ def test_symmetrizing():
     # They should be approximately equal, so check ratios of each one to the next one
     for ind in range(len(cftest)-1):
         np.testing.assert_almost_equal(
-            (cftest[ind]/cftest[ind+1] - 1.)/symm_divide, 0., decimal=decimal_approx,
+            (cftest[ind]/cftest[ind+1] - 1.)/symm_divide, 0., decimal=2,
             err_msg="Noise field generated by symmetrizing COSMOS CorrelatedNoise does not have "+
             "approximate 4-fold symmetry")
 
@@ -799,7 +869,7 @@ def test_symmetrizing():
     cftest00 = cntest_symmetrized._profile.xValue(galsim.PositionD(0., 0.))
     # Test variances first
     np.testing.assert_almost_equal(
-        (cftest00/sym_variance - 1.)/symm_divide, 0., decimal=decimal_approx,
+        (cftest00/sym_variance - 1.)/symm_divide, 0., decimal=2,
         err_msg="Noise field generated by symmetrizing rotated, sheared, magnified COSMOS "+
         "CorrelatedNoise does not approximately match theoretical variance")
     # Then test (1, 0), (0, 1), (-1,0) and (0,-1) values
@@ -811,7 +881,7 @@ def test_symmetrizing():
     # They should be approximately equal, so check ratios of each one to the next one
     for ind in range(len(cftest)-1):
         np.testing.assert_almost_equal(
-            (cftest[ind]/cftest[ind+1]-1.)/symm_divide, 0., decimal=decimal_approx,
+            (cftest[ind]/cftest[ind+1]-1.)/symm_divide, 0., decimal=2,
             err_msg="Noise field generated by symmetrizing rotated, sheared, magnified COSMOS "+
             "CorrelatedNoise does not have approximate 4-fold symmetry")
 
@@ -836,7 +906,7 @@ def test_symmetrizing():
     cftest00 = cntest_symmetrized._profile.xValue(galsim.PositionD(0., 0.))
     # Test variances first
     np.testing.assert_almost_equal(
-        (cftest00/sym_variance - 1.)/symm_divide, 0., decimal=decimal_approx,
+        (cftest00/sym_variance - 1.)/symm_divide, 0., decimal=2,
         err_msg="Noise field generated by symmetrizing rotated, sheared, magnified, convolved "+
         "COSMOS CorrelatedNoise does not approximately match theoretical variance")
     # Then test (1, 0), (0, 1), (-1, 0) and (0, -1) values
@@ -848,7 +918,7 @@ def test_symmetrizing():
     # They should be approximately equal, so check ratios of each one to the next one
     for ind in range(len(cftest)-1):
         np.testing.assert_almost_equal(
-            (cftest[ind]/cftest[ind+1] - 1.)/symm_divide, 0., decimal=decimal_approx,
+            (cftest[ind]/cftest[ind+1] - 1.)/symm_divide, 0., decimal=2,
             err_msg="Noise field generated by symmetrizing rotated, sheared, magnified, convolved "+
             "COSMOS CorrelatedNoise does not have approximate 4-fold symmetry")
     t2 = time.time()
@@ -878,7 +948,8 @@ def test_convolve_cosmos():
     # Then define the convolved cosmos correlated noise model
     conv_cn = cn.convolvedWith(psf_shera)
     # Then draw the correlation function for this correlated noise as the reference
-    refim = galsim.ImageD(smallim_size, smallim_size)
+    # Note: non-square and odd size for y
+    refim = galsim.ImageD(smallim_size, smallim_size + 11)
     conv_cn.drawImage(refim, scale=0.18)
     # Now start the test...
     # First we generate a COSMOS noise field (cosimage), read it into an InterpolatedImage and
@@ -887,7 +958,7 @@ def test_convolve_cosmos():
     # Number of tests needs to be a little larger to beat down noise here, but see the script
     # in devel/external/test_cf/test_cf_convolution_detailed.py
     cosimage_padded = galsim.ImageD(
-        (2 * smallim_size) * 6 + 256, # Note 6 here since 0.18 = 6 * 0.03
+        (2 * smallim_size) * 6 + 355, # Note 6 here since 0.18 = 6 * 0.03
         (2 * smallim_size) * 6 + 256, # large image to beat down noise + padding
         scale=cosmos_scale)           # Use COSMOS pixel scale
     cosimage_padded.addNoise(cn) # Add cosmos noise
@@ -897,21 +968,32 @@ def test_convolve_cosmos():
         normalization='sb', scale=cosmos_scale, x_interpolant=interp)
     cimobj_padded = galsim.Convolve(imobj_padded, psf_shera)
 
+    if __name__ == "__main__":
+        # The convolve_cosmos test, which includes a lot of the correlated noise functionality is
+        # fairly sensitive at 2dp, but takes ~200s on a mid-range laptop
+        decimal_convolve_cosmos = 2
+        nsum_test_convolve_cosmos = 1000
+    else:
+        # Basic settings for convolve_cosmos, will only catch basic screwups
+        decimal_convolve_cosmos = 1
+        nsum_test_convolve_cosmos = 10
+
     # We draw, calculate a correlation function for the resulting field, and repeat to get an
     # average over nsum_test_convolve_cosmos trials
-    convimage = galsim.ImageD(2 * smallim_size, 2 * smallim_size)
-    cimobj_padded.drawImage(convimage, scale=0.18, method='sb')
-    cn_test = galsim.CorrelatedNoise(
-        convimage, gd, scale=0.18, correct_periodicity=True, subtract_mean=False)
-    testim = galsim.ImageD(smallim_size, smallim_size)
+    convimage = galsim.ImageD(2 * smallim_size, 2 * smallim_size + 11)
+    cimobj_padded.draw(convimage, scale=0.18, normalization='sb')
+    cn_test = galsim.CorrelatedNoise(convimage, gd, correct_periodicity=True, subtract_mean=False)
+    testim = galsim.ImageD(smallim_size, smallim_size + 11)
     cn_test.drawImage(testim, scale=0.18)
     # Start some lists to store image info
     conv_list = [convimage.array.copy()] # Don't forget Python reference/assignment semantics, we
                                          # zero convimage and write over it later!
     mnsq_list = [np.mean(convimage.array**2)]
     var_list = [convimage.array.var()]
-    for i in range(nsum_test_convolve_cosmos - 1): # See notes for nsum_test_convolve_cosmos
-                                                   # at the top of the module: varies test accuracy
+    print 'start set of {} iterations to build up the correlation function'.format(
+            nsum_test_convolve_cosmos)
+    for i in range(nsum_test_convolve_cosmos - 1):
+        print 'iteration ',i
         cosimage_padded.setZero()
         cosimage_padded.addNoise(cn)
         imobj_padded = galsim.InterpolatedImage(
@@ -924,8 +1006,8 @@ def test_convolve_cosmos():
         conv_list.append(convimage.array.copy()) # See above
         mnsq_list.append(np.mean(convimage.array**2))
         var_list.append(convimage.array.var())
-        cn_test = galsim.CorrelatedNoise(
-            convimage, gd, scale=0.18, correct_periodicity=True, subtract_mean=False) 
+        cn_test = galsim.CorrelatedNoise(convimage, gd, correct_periodicity=True,
+                                         subtract_mean=False) 
         cn_test.drawImage(testim, scale=0.18, add_to_image=True)
         del imobj_padded
         del cimobj_padded
@@ -973,12 +1055,12 @@ def test_uncorrelated_noise_tracking():
 
     # Start with an UncorrelatedNoise instance that we attach to an InterpolatedImage GSObject as a
     # 'noise' attribute
-    gal_sigma = 1.
-    noise_var = 1.
+    gal_sigma = 1.7
+    noise_var = 1.3
     seed = 1234
     pix_scale = 0.1
     orig_object = galsim.Gaussian(sigma=gal_sigma)
-    orig_ucn = galsim.UncorrelatedNoise(noise_var, rng=galsim.BaseDeviate(seed), scale=pix_scale)
+    orig_ucn = galsim.UncorrelatedNoise(variance=noise_var, scale=pix_scale)
     im = orig_object.drawImage(scale=pix_scale, method='no_pixel')
     int_im = galsim.InterpolatedImage(im)
     # Note, I'm including the noise attribute without actually adding noise.  It doesn't matter
@@ -1000,9 +1082,9 @@ def test_uncorrelated_noise_tracking():
 
     # Now, make a correlated noise object directly based on a realization of the original
     # uncorrelated noise object.
-    test_im = galsim.Image(512,512)
+    test_im = galsim.Image(512,512, scale=pix_scale)
     orig_ucn.applyTo(test_im)
-    cn = galsim.CorrelatedNoise(test_im, galsim.BaseDeviate(seed), scale=pix_scale)
+    cn = galsim.CorrelatedNoise(test_im, galsim.BaseDeviate(seed))
 
     # Run it through the same operations.
     new_cn = cn.shear(g1=test_shear)
@@ -1015,8 +1097,9 @@ def test_uncorrelated_noise_tracking():
     # Allow for some error due to inferring the CorrelatedNoise object 'cn' from a single
     # realization.  For now we'll do the simplest possible comparison of just the variance.  This is
     # probably not adequate but it's a start.
-    np.testing.assert_almost_equal(final_noise.getVariance(), new_cn.getVariance(), decimal=3,
-                                   err_msg='Failure in tracking noise properties through operations')
+    np.testing.assert_almost_equal(
+            final_noise.getVariance(), new_cn.getVariance(), decimal=3,
+            err_msg='Failure in tracking noise properties through operations')
 
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
@@ -1049,9 +1132,84 @@ def test_variance_changes():
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(), t2 - t1)
 
+def test_cosmos_wcs():
+    """Test how getCOSMOSNoise works when applied to an image with a WCS.
+    """
+    t1 = time.time()
+
+    var = 1.7
+    rng = galsim.BaseDeviate(8675309)
+    cn_cosmos = galsim.getCOSMOSNoise(rng=rng, variance=var)
+    cosmos_scale = 0.03
+
+    # Shear it significantly to amplify the directionality of the correlations.
+    cn_orig = cn_cosmos.shear(e2=0.7)
+
+    test_wcs_list = [
+        galsim.PixelScale(cosmos_scale),                            # Same as original
+        #galsim.PixelScale(2.*cosmos_scale),                         # 2x larger pixels
+        galsim.PixelScale(0.5*cosmos_scale),                        # 2x smaller pixels
+        galsim.JacobianWCS(0., cosmos_scale, -cosmos_scale, 0.),    # 90 degrees rotated
+    ]
+    # The test doesn't work for the big pixels.  When drawing on such an image, there
+    # is too much information lost by sampling on the effectively larger grid, so the
+    # measured correlation function is unable to recover the original noise correlations
+    # correctly.  I don't think this is a problem, so don't include that in the test list.
+
+    # This test isn't super fast, so for regular unit tests, just do the last one.
+    if __name__ != "__main__":
+        test_wcs_list = test_wcs_list[-1:]
+
+    for k, test_wcs in enumerate(test_wcs_list):
+        print test_wcs
+        test_im = galsim.ImageD(3 * largeim_size, 3 * largeim_size, wcs=test_wcs)
+
+        # This adds the noise respecting the different WCS functions in the two cases.
+        test_im.addNoise(cn_orig)
+        cn_test = galsim.CorrelatedNoise(test_im)
+
+        # The "raw" correlation function treats the image as having the cosmos pixel_scale
+        cn_raw = galsim.CorrelatedNoise(test_im.view(scale=cosmos_scale))
+
+        # Check basic correlation function values of the 3x3 pixel region around (0,0)
+        for xpos, ypos in zip((0., cosmos_scale, 0., cosmos_scale, cosmos_scale), 
+                              (0., 0., cosmos_scale, -cosmos_scale, cosmos_scale)):
+            pos = galsim.PositionD(xpos, ypos)
+            cf_orig = cn_orig._profile.xValue(pos)
+            cf_test = cn_test._profile.xValue(pos)
+            cf_raw = cn_raw._profile.xValue(pos)
+            print pos, cf_orig, cf_test, cf_raw
+            np.testing.assert_almost_equal(
+                    cf_orig/var, cf_test/var, decimal=2,
+                    err_msg='Drawing COSMOS noise on image with WCS did not ' +
+                    'recover correct covariance at positions '+str(pos))
+
+        # Repeat, but this time adding the noise to a view with the cosmos pixel scale
+        test_im.setZero()
+        test_im.view(wcs=cn_orig.wcs).addNoise(cn_orig)
+        cn_test = galsim.CorrelatedNoise(test_im)
+        cn_raw = galsim.CorrelatedNoise(test_im.view(scale=cosmos_scale))
+
+        # This time it is the raw cf values that should match.
+        for xpos, ypos in zip((0., cosmos_scale, 0., cosmos_scale, cosmos_scale), 
+                              (0., 0., cosmos_scale, -cosmos_scale, cosmos_scale)):
+            pos = galsim.PositionD(xpos, ypos)
+            cf_orig = cn_orig._profile.xValue(pos)
+            cf_test = cn_test._profile.xValue(pos)
+            cf_raw = cn_raw._profile.xValue(pos)
+            print pos, cf_orig, cf_test, cf_raw
+            np.testing.assert_almost_equal(
+                    cf_orig/var, cf_raw/var, decimal=2,
+                    err_msg='Drawing COSMOS noise on view with cosmos pixel scale did not '+
+                    'recover correct covariance at positions '+str(pos))
+
+    t2 = time.time()
+    print 'time for %s = %.2f'%(funcname(), t2 - t1)
+
 if __name__ == "__main__":
     test_uncorrelated_noise_zero_lag()
     test_uncorrelated_noise_nonzero_lag()
+    test_uncorrelated_noise_output()
     test_uncorrelated_noise_symmetry_90degree_rotation()
     test_xcorr_noise_basics_symmetry_90degree_rotation()
     test_ycorr_noise_basics_symmetry_90degree_rotation()
@@ -1068,4 +1226,5 @@ if __name__ == "__main__":
     test_convolve_cosmos()
     test_uncorrelated_noise_tracking()
     test_variance_changes()
+    test_cosmos_wcs()